refactor(openshell): isolate bounded asynchronous capture
